Refactored tests
There was a very confusing idiom for verifying whether Derby was up or
not: Assert.assertFalse(isDerbyUp(startMojo)),
where a false value for isDerbyUp meant that Derby was up and an
SQLException being thrown meant that it was not.

This refactoring focuses the tests on normal vs. skipped execution.

It was necessary because I am going to introduce additional ways to
configure skipping of tests. 
The current style was not well suited for adding more tests for skipped
execution.